+++ internal/usbgadget/recovery.go
@@ -1,0 +1,16 @@
+package usbgadget
+
+import "time"
+
+// USBRecoveryRetryInterval is the minimum interval between USB recovery attempts.
+const USBRecoveryRetryInterval = 5 * time.Second
+
+// ShouldAttemptUSBRecovery returns true if a USB gadget recovery should be attempted,
+// based on the current USB state, whether emulation is desired, and rate limiting.
+func ShouldAttemptUSBRecovery(state string, desired bool, lastAttempt time.Time, now time.Time) bool {
+	if state != "not attached" || !desired {
+		return false
+	}
+
+	return lastAttempt.IsZero() || now.Sub(lastAttempt) >= USBRecoveryRetryInterval
+}
